requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.10.4
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
black==23.3.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.92.0
fastapi-cache2==0.2.1
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
iso8601==1.1.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
orjson==3.8.11
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pendulum==2.1.2
platformdirs==3.5.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.5
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.5.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.25.0
tomli==2.0.1
tornado==6.3.1
twitchio==2.6.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
yarl==1.9.2
yt-dlp==2023.3.4
